lgtm [java/integer-multiplication-cast-to-long]

+++ RoaringBitmap/src/main/java/org/roaringbitmap/bsi/RoaringBitmapSliceIndex.java
@@ -448,8 +448,7 @@ public class RoaringBitmapSliceIndex implements BitmapSliceIndex {
         long count = foundSet.getLongCardinality();
 
         Long sum = IntStream.range(0, this.bitCount())
-                .mapToLong(x -> (1 << x) * RoaringBitmap.andCardinality(this.bA[x], foundSet))
-//        lgtm [java/integer-multiplication-cast-to-long]
+                .mapToLong(x -> (long) (1 << x) * RoaringBitmap.andCardinality(this.bA[x], foundSet))
                 .sum();
 
         return Pair.newPair(sum, count);
