+++ RoaringBitmap/src/main/java/org/roaringbitmap/bsi/buffer/BitSliceIndexBase.java
@@ -441,8 +441,7 @@ public class BitSliceIndexBase {
         long count = foundSet.getLongCardinality();
 
         Long sum = IntStream.range(0, this.bitCount())
-                .mapToLong(x -> (1 << x) * ImmutableRoaringBitmap.andCardinality(this.bA[x], foundSet))
-//        lgtm [java/integer-multiplication-cast-to-long]
+                .mapToLong(x -> (long) (1 << x) * ImmutableRoaringBitmap.andCardinality(this.bA[x], foundSet))
                 .sum();
 
         return Pair.newPair(sum, count);
